WCS 2.0 working
