+++ lib/Models/Catalog/Ows/WebMapServiceCatalogItem.ts
@@ -10,811 +10,48 @@
 import i18next from "i18next";
 import { computed, runInAction } from "mobx";
 import combine from "terriajs-cesium/Source/Core/combine";
-import Ellipsoid from "terriajs-cesium/Source/Core/Ellipsoid";
 import GeographicTilingScheme from "terriajs-cesium/Source/Core/GeographicTilingScheme";
 import JulianDate from "terriajs-cesium/Source/Core/JulianDate";
 import WebMercatorTilingScheme from "terriajs-cesium/Source/Core/WebMercatorTilingScheme";
-import ImageryProvider from "terriajs-cesium/Source/Scene/ImageryProvider";
 import WebMapServiceImageryProvider from "terriajs-cesium/Source/Scene/WebMapServiceImageryProvider";
 import URI from "urijs";
-import containsAny from "../../../Core/containsAny";
-import createDiscreteTimesFromIsoSegments from "../../../Core/createDiscreteTimes";
 import createTransformerAllowUndefined from "../../../Core/createTransformerAllowUndefined";
 import filterOutUndefined from "../../../Core/filterOutUndefined";
 import isDefined from "../../../Core/isDefined";
-import isReadOnlyArray from "../../../Core/isReadOnlyArray";
-import { JsonObject } from "../../../Core/Json";
-import TerriaError from "../../../Core/TerriaError";
 import CatalogMemberMixin from "../../../ModelMixins/CatalogMemberMixin";
 import ChartableMixin from "../../../ModelMixins/ChartableMixin";
 import DiffableMixin from "../../../ModelMixins/DiffableMixin";
 import ExportableMixin from "../../../ModelMixins/ExportableMixin";
 import GetCapabilitiesMixin from "../../../ModelMixins/GetCapabilitiesMixin";
 import { ImageryParts } from "../../../ModelMixins/MappableMixin";
+import MinMaxLevelMixin from "../../../ModelMixins/MinMaxLevelMixin";
 import TileErrorHandlerMixin from "../../../ModelMixins/TileErrorHandlerMixin";
-import TimeFilterMixin from "../../../ModelMixins/TimeFilterMixin";
 import UrlMixin from "../../../ModelMixins/UrlMixin";
+import WebMapServiceCatalogItemTraits, {
+  SUPPORTED_CRS_3857,
+  SUPPORTED_CRS_4326
+} from "../../../Traits/TraitsClasses/WebMapServiceCatalogItemTraits";
+import CommonStrata from "../../Definition/CommonStrata";
+import CreateModel from "../../Definition/CreateModel";
 import SelectableDimensions, {
   SelectableDimension,
   SelectableDimensionSelect
 } from "../../SelectableDimensions";
-import { terriaTheme } from "../../../ReactViews/StandardUserInterface/StandardTheme";
-import {
-  InfoSectionTraits,
-  MetadataUrlTraits
-} from "../../../Traits/TraitsClasses/CatalogMemberTraits";
-import LegendTraits from "../../../Traits/TraitsClasses/LegendTraits";
-import { RectangleTraits } from "../../../Traits/TraitsClasses/MappableTraits";
-import WebMapServiceCatalogItemTraits, {
-  SUPPORTED_CRS_3857,
-  SUPPORTED_CRS_4326,
-  WebMapServiceAvailableLayerDimensionsTraits,
-  WebMapServiceAvailableLayerStylesTraits,
-  WebMapServiceAvailableStyleTraits
-} from "../../../Traits/TraitsClasses/WebMapServiceCatalogItemTraits";
-import { callWebCoverageService } from "./callWebCoverageService";
-import CommonStrata from "../../Definition/CommonStrata";
-import CreateModel from "../../Definition/CreateModel";
-import createStratumInstance from "../../Definition/createStratumInstance";
-import LoadableStratum from "../../Definition/LoadableStratum";
-import Model, { BaseModel } from "../../Definition/Model";
-import { CapabilitiesStyle } from "./OwsInterfaces";
 import proxyCatalogItemUrl from "../proxyCatalogItemUrl";
-import StratumFromTraits from "../../Definition/StratumFromTraits";
-import WebMapServiceCapabilities, {
-  CapabilitiesContactInformation,
-  CapabilitiesDimension,
-  CapabilitiesLayer,
-  getRectangleFromLayer,
-  MetadataURL
-} from "./WebMapServiceCapabilities";
+import { callWebCoverageService } from "./callWebCoverageService";
+import WebMapServiceCapabilities from "./WebMapServiceCapabilities";
+import WebMapServiceCapabilitiesStratum from "./WebMapServiceCapabilitiesStratum";
 import WebMapServiceCatalogGroup from "./WebMapServiceCatalogGroup";
-import MinMaxLevelMixin from "../../../ModelMixins/MinMaxLevelMixin";
-
-const dateFormat = require("dateformat");
-class GetCapabilitiesStratum extends LoadableStratum(
-  WebMapServiceCatalogItemTraits
-) {
-  static async load(
-    catalogItem: WebMapServiceCatalogItem,
-    capabilities?: WebMapServiceCapabilities
-  ): Promise<GetCapabilitiesStratum> {
-    if (!isDefined(catalogItem.getCapabilitiesUrl)) {
-      throw new TerriaError({
-        title: i18next.t("models.webMapServiceCatalogItem.missingUrlTitle"),
-        message: i18next.t("models.webMapServiceCatalogItem.missingUrlMessage")
-      });
-    }
-
-    if (!isDefined(capabilities))
-      capabilities = await WebMapServiceCapabilities.fromUrl(
-        proxyCatalogItemUrl(
-          catalogItem,
-          catalogItem.getCapabilitiesUrl,
-          catalogItem.getCapabilitiesCacheDuration
-        )
-      );
-
-    return new GetCapabilitiesStratum(catalogItem, capabilities);
-  }
-
-  constructor(
-    readonly catalogItem: WebMapServiceCatalogItem,
-    readonly capabilities: WebMapServiceCapabilities
-  ) {
-    super();
-  }
-
-  duplicateLoadableStratum(model: BaseModel): this {
-    return new GetCapabilitiesStratum(
-      model as WebMapServiceCatalogItem,
-      this.capabilities
-    ) as this;
-  }
-
-  @computed get metadataUrls() {
-    const metadataUrls: MetadataURL[] = [];
-
-    Array.from(this.capabilitiesLayers.values()).forEach(layer => {
-      if (!layer?.MetadataURL) return;
-      Array.isArray(layer?.MetadataURL)
-        ? metadataUrls.push(...layer?.MetadataURL)
-        : metadataUrls.push(layer?.MetadataURL as MetadataURL);
-    });
-
-    return metadataUrls
-      .filter(m => m.OnlineResource?.["xlink:href"])
-      .map(m =>
-        createStratumInstance(MetadataUrlTraits, {
-          url: m.OnlineResource!["xlink:href"]
-        })
-      );
-  }
-
-  @computed
-  get layers(): string | undefined {
-    let layers: string | undefined;
-
-    if (this.catalogItem.uri !== undefined) {
-      // Try to extract a layer from the URL
-      const query: any = this.catalogItem.uri.query(true);
-      layers = query.layers;
-    }
-
-    if (layers === undefined) {
-      // Use all the top-level, named layers
-      layers = filterOutUndefined(
-        this.capabilities.topLevelNamedLayers.map(layer => layer.Name)
-      ).join(",");
-    }
-
-    return layers;
-  }
-
-  /**
- * **How we determine WMS legends (in order)**
-  1. Defined manually in catalog JSON
-  2. If `style` is undefined, and server doesn't support `GetLegendGraphic`, we must select first style as default - as there is no way to know what the default style is, and to request a legend for it
-  3. If `style` is is set and it has a `legendUrl` -> use it!
-  4. If server supports `GetLegendGraphic`, we can request a legend (with or without `style` parameter)
- */
-  @computed
-  get legends(): StratumFromTraits<LegendTraits>[] | undefined {
-    const availableStyles = this.catalogItem.availableStyles || [];
-    const layers = this.catalogItem.layersArray;
-    const styles = this.catalogItem.stylesArray;
-
-    const result: StratumFromTraits<LegendTraits>[] = [];
-
-    for (let i = 0; i < layers.length; ++i) {
-      const layer = layers[i];
-      const style = i < styles.length ? styles[i] : undefined;
-
-      let legendUri: uri.URI | undefined;
-      let legendUrlMimeType: string | undefined;
-      let legendScaling: number | undefined;
-
-      const layerAvailableStyles = availableStyles.find(
-        candidate => candidate.layerName === layer
-      )?.styles;
-
-      let layerStyle: Model<WebMapServiceAvailableStyleTraits> | undefined;
-
-      if (isDefined(style)) {
-        // Attempt to find layer style based on AvailableStyleTraits
-        layerStyle = layerAvailableStyles?.find(
-          candidate => candidate.name === style
-        );
-      }
-
-      // If no style is selected and this WMS doesn't support GetLegendGraphics - we must use the first style if none is explicitly specified.
-      // (If WMS supports GetLegendGraphics we can use it and omit style parameter to get the "default" style's legend)
-      if (!isDefined(layerStyle) && !this.catalogItem.supportsGetLegendGraphic)
-        layerStyle = layerAvailableStyles?.[0];
-
-      // If legend found - proxy URL and set mimetype
-      if (layerStyle?.legend?.url) {
-        legendUri = URI(
-          proxyCatalogItemUrl(this.catalogItem, layerStyle.legend.url)
-        );
-
-        legendUrlMimeType = layerStyle.legend.urlMimeType;
-      }
-
-      // If no legends found and WMS supports GetLegendGraphics - make one up!
-      if (
-        !isDefined(legendUri) &&
-        isDefined(this.catalogItem.url) &&
-        this.catalogItem.supportsGetLegendGraphic
-      ) {
-        legendUri = URI(
-          proxyCatalogItemUrl(
-            this.catalogItem,
-            this.catalogItem.url.split("?")[0]
-          )
-        );
-        legendUri
-          .setQuery("service", "WMS")
-          .setQuery("version", "1.3.0")
-          .setQuery("request", "GetLegendGraphic")
-          .setQuery("format", "image/png")
-          .setQuery("layer", layer);
-
-        // From OGC — about style property for GetLegendGraphic request:
-        // If not present, the default style is selected. The style may be any valid style available for a layer, including non-SLD internally-defined styles.
-        if (style) {
-          legendUri.setQuery("style", style);
-        }
-        legendUrlMimeType = "image/png";
-      }
-
-      if (isDefined(legendUri)) {
-        // Add geoserver related LEGEND_OPTIONS to match terria styling (if supported)
-        if (
-          this.catalogItem.isGeoServer &&
-          legendUri.hasQuery("request", "GetLegendGraphic")
-        ) {
-          let legendOptions =
-            "fontName:Courier;fontStyle:bold;fontSize:12;forceLabels:on;fontAntiAliasing:true;labelMargin:5";
-
-          // Geoserver fontColor must be a hex value - use `textLight` theme colour
-          let fontColor = terriaTheme.textLight.split("#")?.[1];
-          if (isDefined(fontColor)) {
-            // If fontColor is a 3-character hex -> turn into 6
-            if (fontColor.length === 3) {
-              fontColor = `${fontColor[0]}${fontColor[0]}${fontColor[1]}${fontColor[1]}${fontColor[2]}${fontColor[2]}`;
-            }
-            legendOptions += `;fontColor:0x${fontColor}`;
-          }
-
-          legendOptions += ";dpi:182"; // enable if we can scale the image back down by 50%.
-          legendScaling = 0.5;
-          legendUri.setQuery("LEGEND_OPTIONS", legendOptions);
-          legendUri.setQuery("transparent", "true");
-        }
-
-        // Add colour scale range params if supported
-        if (
-          this.catalogItem.supportsColorScaleRange &&
-          this.catalogItem.colorScaleRange
-        ) {
-          legendUri.setQuery(
-            "colorscalerange",
-            this.catalogItem.colorScaleRange
-          );
-        }
-        result.push(
-          createStratumInstance(LegendTraits, {
-            url: legendUri.toString(),
-            urlMimeType: legendUrlMimeType,
-            imageScaling: legendScaling
-          })
-        );
-      }
-    }
-
-    return result;
-  }
-
-  @computed
-  get capabilitiesLayers(): ReadonlyMap<string, CapabilitiesLayer | undefined> {
-    const lookup: (
-      name: string
-    ) => [string, CapabilitiesLayer | undefined] = name => [
-      name,
-      this.capabilities && this.capabilities.findLayer(name)
-    ];
-    return new Map(this.catalogItem.layersArray.map(lookup));
-  }
-
-  @computed get crs() {
-    // Get set of supported CRS from layer hierarchy
-    const layerCrs = new Set<string>();
-    this.capabilitiesLayers.forEach(layer => {
-      if (layer) {
-        const srs = this.capabilities.getInheritedValues(layer, "SRS");
-        const crs = this.capabilities.getInheritedValues(layer, "CRS");
-        [
-          ...(Array.isArray(srs) ? srs : [srs]),
-          ...(Array.isArray(crs) ? crs : [crs])
-        ].forEach(c => layerCrs.add(c));
-      }
-    });
-
-    // Note order is important here, the first one found will be used
-    const supportedCrs = [...SUPPORTED_CRS_3857, ...SUPPORTED_CRS_4326];
-
-    // If nothing is supported, ask for EPSG:3857, and hope for the best.
-    return supportedCrs.find(crs => layerCrs.has(crs)) ?? "EPSG:3857";
-  }
-
-  @computed
-  get availableDimensions(): StratumFromTraits<
-    WebMapServiceAvailableLayerDimensionsTraits
-  >[] {
-    const result: StratumFromTraits<
-      WebMapServiceAvailableLayerDimensionsTraits
-    >[] = [];
-
-    if (!this.capabilities) {
-      return result;
-    }
-
-    const capabilitiesLayers = this.capabilitiesLayers;
-
-    for (const layerTuple of capabilitiesLayers) {
-      const layerName = layerTuple[0];
-      const layer = layerTuple[1];
-
-      const dimensions: ReadonlyArray<CapabilitiesDimension> = layer
-        ? this.capabilities.getInheritedValues(layer, "Dimension")
-        : [];
-
-      result.push({
-        layerName: layerName,
-        dimensions: dimensions
-          .filter(dim => dim.name !== "time")
-          .map(dim => {
-            return {
-              name: dim.name,
-              units: dim.units,
-              unitSymbol: dim.unitSymbol,
-              default: dim.default,
-              multipleValues: dim.multipleValues,
-              current: dim.current,
-              nearestValue: dim.nearestValue,
-              values: dim.text?.split(",")
-            };
-          })
-      });
-    }
-
-    return result;
-  }
-
-  @computed
-  get availableStyles(): StratumFromTraits<
-    WebMapServiceAvailableLayerStylesTraits
-  >[] {
-    const result: StratumFromTraits<
-      WebMapServiceAvailableLayerStylesTraits
-    >[] = [];
-
-    if (!this.capabilities) {
-      return result;
-    }
-
-    const capabilitiesLayers = this.capabilitiesLayers;
-    for (const layerTuple of capabilitiesLayers) {
-      const layerName = layerTuple[0];
-      const layer = layerTuple[1];
-
-      const styles: ReadonlyArray<CapabilitiesStyle> = layer
-        ? this.capabilities.getInheritedValues(layer, "Style")
-        : [];
-      result.push({
-        layerName: layerName,
-        styles: styles.map(style => {
-          var wmsLegendUrl = isReadOnlyArray(style.LegendURL)
-            ? style.LegendURL[0]
-            : style.LegendURL;
-
-          var legendUri, legendMimeType;
-          if (
-            wmsLegendUrl &&
-            wmsLegendUrl.OnlineResource &&
-            wmsLegendUrl.OnlineResource["xlink:href"]
-          ) {
-            legendUri = new URI(
-              decodeURIComponent(wmsLegendUrl.OnlineResource["xlink:href"])
-            );
-            legendMimeType = wmsLegendUrl.Format;
-          }
-          const legend = !legendUri
-            ? undefined
-            : createStratumInstance(LegendTraits, {
-                url: legendUri.toString(),
-                urlMimeType: legendMimeType,
-                title:
-                  (capabilitiesLayers.size > 1 && layer?.Title) || undefined // Add layer Title as legend title if showing multiple layers
-              });
-
-          return {
-            name: style.Name,
-            title: style.Title,
-            abstract: style.Abstract,
-            legend: legend
-          };
-        })
-      });
-    }
-
-    return result;
-  }
-
-  @computed
-  get info(): StratumFromTraits<InfoSectionTraits>[] {
-    const result: StratumFromTraits<InfoSectionTraits>[] = [];
-
-    let firstDataDescription: string | undefined;
-
-    result.push(
-      createStratumInstance(InfoSectionTraits, {
-        name: i18next.t("models.webMapServiceCatalogItem.serviceDescription"),
-        contentAsObject: this.capabilities.Service as JsonObject,
-        // Hide big ugly table by default
-        show: false
-      })
-    );
-
-    const onlyHasSingleLayer = this.catalogItem.layersArray.length === 1;
-
-    if (onlyHasSingleLayer) {
-      // Clone the capabilitiesLayer as we'll modify it in a second
-      const out = Object.assign(
-        {},
-        this.capabilitiesLayers.get(this.catalogItem.layersArray[0])
-      ) as any;
-
-      if (out !== undefined) {
-        // The Dimension object is really weird and has a bunch of stray text in there
-        if ("Dimension" in out) {
-          const goodDimension: any = {};
-          Object.keys(out.Dimension).forEach((k: any) => {
-            if (isNaN(k)) {
-              goodDimension[k] = out.Dimension[k];
-            }
-          });
-          out.Dimension = goodDimension;
-        }
-
-        // remove a circular reference to the parent
-        delete out._parent;
-
-        try {
-          result.push(
-            createStratumInstance(InfoSectionTraits, {
-              name: i18next.t(
-                "models.webMapServiceCatalogItem.dataDescription"
-              ),
-              contentAsObject: out as JsonObject,
-              // Hide big ugly table by default
-              show: false
-            })
-          );
-        } catch (e) {
-          console.log(
-            `FAILED to create InfoSection with WMS layer Capabilities`
-          );
-          console.log(e);
-        }
-      }
-    }
-
-    for (const layer of this.capabilitiesLayers.values()) {
-      if (
-        !layer ||
-        !layer.Abstract ||
-        containsAny(layer.Abstract, WebMapServiceCatalogItem.abstractsToIgnore)
-      ) {
-        continue;
-      }
-
-      const suffix =
-        this.capabilitiesLayers.size === 1 ? "" : ` - ${layer.Title}`;
-      const name = `Web Map Service Layer Description${suffix}`;
-      result.push(
-        createStratumInstance(InfoSectionTraits, {
-          name,
-          content: layer.Abstract
-        })
-      );
-      firstDataDescription = firstDataDescription || layer.Abstract;
-    }
-
-    // Show the service abstract if there is one and if it isn't the Geoserver default "A compliant implementation..."
-    const service = this.capabilities && this.capabilities.Service;
-    if (service) {
-      if (service.ContactInformation !== undefined) {
-        result.push(
-          createStratumInstance(InfoSectionTraits, {
-            name: i18next.t("models.webMapServiceCatalogItem.serviceContact"),
-            content: getServiceContactInformation(service.ContactInformation)
-          })
-        );
-      }
-
-      result.push(
-        createStratumInstance(InfoSectionTraits, {
-          name: i18next.t("models.webMapServiceCatalogItem.getCapabilitiesUrl"),
-          content: this.catalogItem.getCapabilitiesUrl
-        })
-      );
-
-      if (
-        service &&
-        service.Abstract &&
-        !containsAny(
-          service.Abstract,
-          WebMapServiceCatalogItem.abstractsToIgnore
-        ) &&
-        service.Abstract !== firstDataDescription
-      ) {
-        result.push(
-          createStratumInstance(InfoSectionTraits, {
-            name: i18next.t(
-              "models.webMapServiceCatalogItem.serviceDescription"
-            ),
-            content: service.Abstract
-          })
-        );
-      }
-
-      // Show the Access Constraints if it isn't "none" (because that's the default, and usually a lie).
-      if (
-        service.AccessConstraints &&
-        !/^none$/i.test(service.AccessConstraints)
-      ) {
-        result.push(
-          createStratumInstance(InfoSectionTraits, {
-            name: i18next.t(
-              "models.webMapServiceCatalogItem.accessConstraints"
-            ),
-            content: service.AccessConstraints
-          })
-        );
-      }
-    }
-
-    return result;
-  }
-
-  @computed
-  get infoSectionOrder(): string[] {
-    let layerDescriptions = [`Web Map Service Layer Description`];
-
-    // If more than one layer, push layer description titles for each applicable layer
-    if (this.capabilitiesLayers.size > 1) {
-      layerDescriptions = [];
-      this.capabilitiesLayers.forEach(layer => {
-        if (
-          layer &&
-          layer.Abstract &&
-          !containsAny(
-            layer.Abstract,
-            WebMapServiceCatalogItem.abstractsToIgnore
-          )
-        ) {
-          layerDescriptions.push(
-            `Web Map Service Layer Description - ${layer.Title}`
-          );
-        }
-      });
-    }
-
-    return [
-      i18next.t("preview.disclaimer"),
-      i18next.t("description.name"),
-      ...layerDescriptions,
-      i18next.t("preview.datasetDescription"),
-      i18next.t("preview.serviceDescription"),
-      i18next.t("models.webMapServiceCatalogItem.serviceDescription"),
-      i18next.t("preview.resourceDescription"),
-      i18next.t("preview.licence"),
-      i18next.t("preview.accessConstraints"),
-      i18next.t("models.webMapServiceCatalogItem.accessConstraints"),
-      i18next.t("preview.author"),
-      i18next.t("preview.contact"),
-      i18next.t("models.webMapServiceCatalogItem.serviceContact"),
-      i18next.t("preview.created"),
-      i18next.t("preview.modified"),
-      i18next.t("preview.updateFrequency"),
-      i18next.t("models.webMapServiceCatalogItem.getCapabilitiesUrl")
-    ];
-  }
-
-  @computed
-  get shortReport() {
-    const catalogItem = this.catalogItem;
-    if (catalogItem.isShowingDiff) {
-      const format = "yyyy/mm/dd";
-      const d1 = dateFormat(catalogItem.firstDiffDate, format);
-      const d2 = dateFormat(catalogItem.secondDiffDate, format);
-      return `Showing difference image computed for ${catalogItem.diffStyleId} style on dates ${d1} and ${d2}`;
-    }
-  }
-
-  @computed
-  get rectangle(): StratumFromTraits<RectangleTraits> | undefined {
-    const layers: CapabilitiesLayer[] = [...this.capabilitiesLayers.values()]
-      .filter(layer => layer !== undefined)
-      .map(l => l!);
-    // Needs to take union of all layer rectangles
-    return layers.length > 0 ? getRectangleFromLayer(layers[0]) : undefined;
-    // if (layers.length === 1) {
-    //     return getRectangleFromLayer(layers[0]);
-    // }
-    // Otherwise get the union of rectangles from all layers
-    // return undefined;
-  }
-
-  @computed
-  get isGeoServer(): boolean | undefined {
-    const keyword = this.capabilities?.Service?.KeywordList?.Keyword;
-    return (
-      (isReadOnlyArray(keyword) && keyword.indexOf("GEOSERVER") >= 0) ||
-      keyword === "GEOSERVER" ||
-      this.catalogItem.url?.toLowerCase().includes("geoserver")
-    );
-  }
-
-  // TODO - There is possibly a better way to do this
-  @computed
-  get isThredds(): boolean {
-    if (
-      this.catalogItem.url &&
-      (this.catalogItem.url.indexOf("thredds") > -1 ||
-        this.catalogItem.url.indexOf("tds") > -1)
-    ) {
-      return true;
-    }
-    return false;
-  }
-
-  // TODO - Geoserver also support NCWMS via a plugin, just need to work out how to detect that
-  @computed
-  get isNcWMS(): boolean {
-    if (this.catalogItem.isThredds) return true;
-    return false;
-  }
-
-  @computed
-  get isEsri(): boolean {
-    if (this.catalogItem.url !== undefined)
-      return this.catalogItem.url.indexOf("MapServer/WMSServer") > -1;
-    return false;
-  }
-
-  @computed
-  get supportsGetLegendGraphic(): boolean {
-    return (
-      isDefined(this.capabilities?.json?.["xmlns:sld"]) ||
-      isDefined(
-        this.capabilities?.json?.Capability?.Request?.GetLegendGraphic
-      ) ||
-      (this.catalogItem.isGeoServer ?? false) ||
-      (this.catalogItem.isNcWMS ?? false)
-    );
-  }
-
-  @computed
-  get supportsColorScaleRange(): boolean {
-    return this.catalogItem.isNcWMS;
-  }
-
-  @computed
-  get discreteTimes(): { time: string; tag: string | undefined }[] | undefined {
-    const result = [];
-
-    for (let layer of this.capabilitiesLayers.values()) {
-      if (!layer) {
-        continue;
-      }
-      const dimensions = this.capabilities.getInheritedValues(
-        layer,
-        "Dimension"
-      );
-
-      const timeDimension = dimensions.find(
-        dimension => dimension.name.toLowerCase() === "time"
-      );
-      if (!timeDimension) {
-        continue;
-      }
-
-      let extent: string = timeDimension;
-
-      // WMS 1.1.1 puts dimension values in an Extent element instead of directly in the Dimension element.
-      const extentElements = this.capabilities.getInheritedValues(
-        layer,
-        "Extent"
-      );
-      const extentElement = extentElements.find(
-        extent => extent.name.toLowerCase() === "time"
-      );
-      if (extentElement) {
-        extent = extentElement;
-      }
-
-      if (!extent || !extent.split) {
-        continue;
-      }
-
-      const values = extent.split(",");
-      for (let i = 0; i < values.length; ++i) {
-        const value = values[i];
-        const isoSegments = value.split("/");
-        if (isoSegments.length === 1) {
-          result.push({
-            time: values[i],
-            tag: undefined
-          });
-        } else {
-          createDiscreteTimesFromIsoSegments(
-            result,
-            isoSegments[0],
-            isoSegments[1],
-            isoSegments[2],
-            this.catalogItem.maxRefreshIntervals
-          );
-        }
-      }
-    }
-
-    return result;
-  }
-
-  @computed get currentTime() {
-    // Get default times for all layers
-    const defaultTimes = filterOutUndefined(
-      Array.from(this.capabilitiesLayers).map(([layerName, layer]) => {
-        if (!layer) return;
-        const dimensions = this.capabilities.getInheritedValues(
-          layer,
-          "Dimension"
-        );
-
-        const timeDimension = dimensions.find(
-          dimension => dimension.name.toLowerCase() === "time"
-        );
-
-        return timeDimension?.default;
-      })
-    );
-
-    // Return first default time
-    return defaultTimes[0];
-  }
-}
-
-class DiffStratum extends LoadableStratum(WebMapServiceCatalogItemTraits) {
-  constructor(readonly catalogItem: WebMapServiceCatalogItem) {
-    super();
-  }
-
-  duplicateLoadableStratum(model: BaseModel): this {
-    return new DiffStratum(model as WebMapServiceCatalogItem) as this;
-  }
-
-  @computed
-  get legends() {
-    if (this.catalogItem.isShowingDiff && this.diffLegendUrl) {
-      const urlMimeType =
-        new URL(this.diffLegendUrl).searchParams.get("format") || undefined;
-      return [
-        createStratumInstance(LegendTraits, {
-          url: this.diffLegendUrl,
-          urlMimeType
-        })
-      ];
-    }
-    return undefined;
-  }
-
-  @computed
-  get diffLegendUrl() {
-    const diffStyleId = this.catalogItem.diffStyleId;
-    const firstDate = this.catalogItem.firstDiffDate;
-    const secondDate = this.catalogItem.secondDiffDate;
-    if (diffStyleId && firstDate && secondDate) {
-      return this.catalogItem.getLegendUrlForStyle(
-        diffStyleId,
-        JulianDate.fromIso8601(firstDate),
-        JulianDate.fromIso8601(secondDate)
-      );
-    }
-    return undefined;
-  }
-
-  @computed
-  get disableDateTimeSelector() {
-    return this.catalogItem.isShowingDiff;
-  }
-}
 
 class WebMapServiceCatalogItem
   extends TileErrorHandlerMixin(
     ExportableMixin(
       DiffableMixin(
-        TimeFilterMixin(
-          ChartableMixin(
-            MinMaxLevelMixin(
-              GetCapabilitiesMixin(
-                UrlMixin(
-                  CatalogMemberMixin(
-                    CreateModel(WebMapServiceCatalogItemTraits)
-                  )
-                )
+        ChartableMixin(
+          MinMaxLevelMixin(
+            GetCapabilitiesMixin(
+              UrlMixin(
+                CatalogMemberMixin(CreateModel(WebMapServiceCatalogItemTraits))
               )
             )
           )
@@ -873,7 +110,10 @@ class WebMapServiceCatalogItem
   async createGetCapabilitiesStratumFromParent(
     capabilities: WebMapServiceCapabilities
   ) {
-    const stratum = await GetCapabilitiesStratum.load(this, capabilities);
+    const stratum = await WebMapServiceCapabilitiesStratum.load(
+      this,
+      capabilities
+    );
     runInAction(() => {
       this.strata.set(GetCapabilitiesMixin.getCapabilitiesStratumName, stratum);
     });
@@ -885,12 +125,9 @@ class WebMapServiceCatalogItem
       undefined
     )
       return;
-    const stratum = await GetCapabilitiesStratum.load(this);
+    const stratum = await WebMapServiceCapabilitiesStratum.load(this);
     runInAction(() => {
       this.strata.set(GetCapabilitiesMixin.getCapabilitiesStratumName, stratum);
-
-      const diffStratum = new DiffStratum(this);
-      this.strata.set(DiffableMixin.diffStratumName, diffStratum);
     });
   }
 
@@ -935,10 +172,10 @@ class WebMapServiceCatalogItem
   @computed
   get discreteTimes() {
     const getCapabilitiesStratum:
-      | GetCapabilitiesStratum
+      | WebMapServiceCapabilitiesStratum
       | undefined = this.strata.get(
       GetCapabilitiesMixin.getCapabilitiesStratumName
-    ) as GetCapabilitiesStratum;
+    ) as WebMapServiceCapabilitiesStratum;
     return getCapabilitiesStratum?.discreteTimes;
   }
 
@@ -1012,10 +249,6 @@ class WebMapServiceCatalogItem
       uri.addQuery("time", time);
     }
     return uri.toString();
-  }
-
-  protected forceLoadMapItems(): Promise<void> {
-    return Promise.resolve();
   }
 
   @computed
@@ -1192,9 +425,11 @@ class WebMapServiceCatalogItem
         new URI(this.url)
       );
 
-      const gcStratum: GetCapabilitiesStratum | undefined = this.strata.get(
+      const gcStratum:
+        | WebMapServiceCapabilitiesStratum
+        | undefined = this.strata.get(
         GetCapabilitiesMixin.getCapabilitiesStratumName
-      ) as GetCapabilitiesStratum;
+      ) as WebMapServiceCapabilitiesStratum;
 
       let lyrs: string[] = [];
       if (this.layers && gcStratum !== undefined) {
@@ -1249,8 +484,9 @@ class WebMapServiceCatalogItem
           layer.layerName &&
           (this.strata.get(
             GetCapabilitiesMixin.getCapabilitiesStratumName
-          ) as GetCapabilitiesStratum).capabilitiesLayers.get(layer.layerName)
-            ?.Title;
+          ) as WebMapServiceCapabilitiesStratum).capabilitiesLayers.get(
+            layer.layerName
+          )?.Title;
 
         name = `${layerTitle ||
           layer.layerName ||
@@ -1404,34 +640,6 @@ export function formatDimensionsForOws(
       },
     {}
   );
-}
-
-function getServiceContactInformation(
-  contactInfo: CapabilitiesContactInformation
-) {
-  const primary = contactInfo.ContactPersonPrimary;
-  let text = "";
-  if (isDefined(primary)) {
-    if (
-      isDefined(primary.ContactOrganization) &&
-      primary.ContactOrganization.length > 0 &&
-      // Geoserver default
-      primary.ContactOrganization !== "The Ancient Geographers"
-    ) {
-      text += primary.ContactOrganization + "<br/>";
-    }
-  }
-
-  if (
-    isDefined(contactInfo.ContactElectronicMailAddress) &&
-    contactInfo.ContactElectronicMailAddress.length > 0 &&
-    // Geoserver default
-    contactInfo.ContactElectronicMailAddress !== "claudius.ptolomaeus@gmail.com"
-  ) {
-    text += `[${contactInfo.ContactElectronicMailAddress}](mailto:${contactInfo.ContactElectronicMailAddress})`;
-  }
-
-  return text;
 }
 
 export default WebMapServiceCatalogItem;
